+++ lasy/propagators/__init__.py
@@ -1,7 +1,7 @@
 from .collins_sfft_propagator import CollinsSFFTPropagator
 from .fresnel_sfft_propagator import FresnelSFFTPropagator
 from .propagator import Propagator
-from .propagators_axiprop import (
+from .axiprop_propagators import (
     MRTFresnelPropagator,
     MRTPropagator,
     XYTFresnelPropagator,
